Factor out some hairier reducers to improve readability of adding content-type after changing body type in the form
